feat(Label)!: remove deprecated props
- remove `requiredLabel` from component
- remove `optionalLabel` from component

neither of these are used, and removing them adds clarity and simplicity
to the component.

Also added an additional story to the component documentation for
clarity.
